+++ src/store/machine/connector/PollConnector.js
@@ -4,19 +4,16 @@
 import crc32 from 'turbo-crc32/crc32'
 
 import BaseConnector, { defaultRequestTimeout } from './BaseConnector.js'
-import { getBoardDefinition } from '../boards.js'
 import { DefaultMachineModel } from '../model.js'
-import { HeaterState, StatusType, isPaused, isPrinting } from '../modelEnums.js'
-import { BeepRequest, Layer, Message, MessageBox, ParsedFileInfo } from '../modelItems.js'
+import { StatusType, isPaused, isPrinting } from '../modelEnums.js'
+import { Layer, Message, ParsedFileInfo } from '../modelItems.js'
 
 import {
 	NetworkError, DisconnectedError, TimeoutError, OperationCancelledError, OperationFailedError,
 	DirectoryNotFoundError, FileNotFoundError, DriveUnmountedError,
-	LoginError, InvalidPasswordError, NoFreeSessionError,
+	LoginError, BadVersionError, InvalidPasswordError, NoFreeSessionError,
 	CodeResponseError, CodeBufferError
 } from '../../../utils/errors.js'
-import { arraySizesDiffer, quickPatch } from '../../../utils/patch.js'
-import { bitmapToArray } from '../../../utils/numbers.js'
 import Path from '../../../utils/path.js'
 import { strToTime, timeToStr } from '../../../utils/time.js'
 
@@ -32,14 +29,15 @@ export default class PollConnector extends BaseConnector {
 
 		switch (response.err) {
 			case 0:
-				if (!window.forceLegacyConnect) {
-					if (response.isEmulated) {
-						throw new OperationFailedError('Cancelling connection attempt because the remote endpoint is emulated');
-					}
-					if (response.apiLevel > 0) {
-						// Don't hide the connection dialog while the full model is being loaded...
-						BaseConnector.setConnectingProgress(0);
-					}
+				if (response.isEmulated) {
+					throw new OperationFailedError('Cancelling connection attempt because the remote endpoint is emulated');
+				}
+				if (response.apiLevel > 0) {
+					// Don't hide the connection dialog while the full model is being loaded...
+					BaseConnector.setConnectingProgress(0);
+				} else {
+					// rr_status requests are no longer supported
+					throw new BadVersionError();
 				}
 				return new PollConnector(hostname, password, response);
 			case 1: throw new InvalidPasswordError();
@@ -239,11 +237,9 @@ export default class PollConnector extends BaseConnector {
 	}
 
 	async reconnect() {
-		// Cancel pending requests and reset the last 'seq' field
-		// so that the last G-code reply is queried again
+		// Cancel pending requests and reset the sequence numbers
 		this.cancelRequests();
-		this.lastStatusResponse = {};
-		this.lastSeq = 0;
+		this.lastSeqs = {};
 
 		// Attempt to reconnect
 		const response = await BaseConnector.request('GET', `${location.protocol}//${this.hostname}${process.env.BASE_URL}rr_connect`, {
@@ -285,526 +281,14 @@ export default class PollConnector extends BaseConnector {
 	}
 
 	updateLoopTimer = null
-	updateLoopCounter = 1
-	lastStatusResponse = {}
-	lastSeq = 0
-
-	pendingCodes = []
-	layers = []
-	currentFileInfo = new ParsedFileInfo()
-	printStats = {}
-	probeType = 0
-	numExtruders = 0
-
-	async updateLoopStatus(requestExtendedStatus = false) {
-		requestExtendedStatus |= this.justConnected ||
-			(this.updateLoopCounter % this.settings.extendedUpdateEvery) === 0 ||
-			(this.verbose && (this.updateLoopCounter % 2) === 0);
-
-		// Decide which type of status update to poll and request it
-		const wasPrinting = ['D', 'S', 'R', 'P', 'M'].indexOf(this.lastStatusResponse.status) !== -1;
-		const wasSimulating = this.lastStatusResponse.status === 'M', wasPaused = this.lastStatusResponse.status === 'P';
-		const statusType = requestExtendedStatus ? 2 : (wasPrinting ? 3 : 1);
-		const response = await this.request('GET', 'rr_status', { type: statusType });
-		const isPrinting = ['D', 'S', 'R', 'P', 'M'].indexOf(response.status) !== -1;
-		const newData = {};
-
-		// Check if an extended status response needs to be polled in case machine parameters have changed
-		if (!requestExtendedStatus && arraySizesDiffer(response, this.lastStatusResponse)) {
-			await this.updateLoopStatus(true);
-			return;
-		}
-
-		// Retrieve job file information if a print has started or set info about the last job if it has finished
-		if (isPrinting) {
-			if (this.justConnected || !wasPrinting) {
-				this.currentFileInfo = await this.getFileInfo();
-				delete this.currentFileInfo.printDuration;
-
-				quickPatch(newData, {
-					job: {
-						file: this.currentFileInfo,
-						layers: []
-					}
-				});
-
-				this.layers = [];
-				this.printStats = {
-					layerHeight: this.currentFileInfo.layerHeight
-				};
-			}
-		} else if (wasPrinting) {
-			quickPatch(newData, {
-				job: {
-					lastFileName: this.currentFileInfo.fileName,
-					lastFileCancelled: wasPaused,
-					lastFileSimulated: wasSimulating
-				}
-			});
-		}
-
-		// Standard Status Response
-		const fanRPMs = (response.sensors.fanRPM instanceof Array) ? response.sensors.fanRPM : [response.sensors.fanRPM];
-		quickPatch(newData, {
-			fans: response.params.fanPercent.map((fanPercent, index) => ({
-				rpm: (index < fanRPMs.length && fanRPMs[index] > 0) ? fanRPMs[index] : -1,
-				requestedValue: fanPercent / 100
-			})),
-			heat: {
-				heaters: response.temps.state.map((state, sensor) => ({ state: this.convertHeaterState(state), sensor }), this)
-			},
-			move: {
-				axes: response.coords.xyz.map((position, drive) => ({
-					drives: [drive],
-					homed: Boolean(response.coords.axesHomed[drive]),
-					machinePosition: (position === 9999) ? null : position,
-					userPosition: (position === 9999) ? null : position
-				})),
-				currentMove: {
-					requestedSpeed: (response.speeds !== undefined) ? response.speeds.requested : null,
-					topSpeed: (response.speeds !== undefined) ? response.speeds.top : null
-				},
-				extruders: response.coords.extr.map((position, extruder) => ({
-					drives: [response.coords.xyz.length + extruder],
-					factor: response.params.extrFactors[extruder] / 100,
-					position: position
-				})),
-				speedFactor: response.params.speedFactor / 100,
-				workplaceNumber: (response.wpl !== undefined) ? response.wpl - 1 : 0
-			},
-			scanner: (response.scanner) ? {
-				progress: response.scanner.progress,
-				status: response.scanner.status
-			} : {},
-			sensors: {
-				analog: response.temps.current.map((lastReading) => ({ lastReading }))
-						.concat(response.temps.extra.map(extra => ({
-							lastReading: (extra.temp === 9999) ? null : extra.temp,
-							name: extra.name
-						}))),
-				probes: (this.probeType !== 0) ? [
-					{
-						value: [response.sensors.probeValue].concat(response.sensors.probeSecondary ? response.sensors.probeSecondary : [])
-					}
-				] : []
-			},
-			state: {
-				atxPower: (response.params.atxPower === -1) ? null : Boolean(response.params.atxPower),
-				currentTool: response.currentTool,
-				status: this.convertStatusLetter(response.status)
-			},
-			tools: response.temps.tools.active.map((active, index) => ({
-				active,
-				standby: response.temps.tools.standby[index]
-			}))
-		});
-		if (newData.move.axes.length >= 3) {
-			newData.move.axes[2].babystep = (response.params.babystep !== undefined) ? response.params.babystep : 0;
-		}
-		if (response.coords.machine) {
-			response.coords.machine.forEach(function(machinePosition, axis) {
-				newData.move.axes[axis].machinePosition = (machinePosition === 9999) ? null : machinePosition;
-			});
-		}
-		if (response.temps.bed && response.temps.bed.heater >= 0 && response.temps.bed.heater < newData.sensors.analog.length) {
-			newData.heat.heaters[response.temps.bed.heater].active = response.temps.bed.active;
-			newData.heat.heaters[response.temps.bed.heater].standby = response.temps.bed.standby;
-		}
-		if (response.temps.chamber && response.temps.chamber.heater >= 0 && response.temps.chamber.heater < newData.sensors.analog.length) {
-			newData.heat.heaters[response.temps.chamber.heater].active = response.temps.chamber.active;
-			newData.heat.heaters[response.temps.chamber.heater].standby = response.temps.chamber.standby;
-		}
-		if (response.temps.cabinet && response.temps.cabinet.heater >= 0 && response.temps.cabinet.heater < newData.sensors.analog.length) {
-			newData.heat.heaters[response.temps.cabinet.heater].active = response.temps.cabinet.active;
-			newData.heat.heaters[response.temps.cabinet.heater].standby = response.temps.cabinet.standby;
-		}
-		this.numExtruders = response.coords.extr.length;
-
-		if (statusType === 2) {
-			// Extended Status Response
-			const axisNames = (response.axisNames !== undefined) ? response.axisNames.split('') : ['X', 'Y', 'Z', 'U', 'V', 'W', 'A', 'B', 'C'];
-			const boardDefinition = getBoardDefinition(this.boardType);
-			this.name = name;
-			this.probeType = response.probe ? response.probe.type : 0;
-
-			quickPatch(newData, {
-				boards: [
-					{
-						firmwareFileName: boardDefinition.firmwareFileName,
-						firmwareName: response.firmwareName,
-						iapFileNameSD: boardDefinition.iapFileNameSD,
-						maxHeaters: boardDefinition.maxHeaters,
-						maxMotors: boardDefinition.maxMotors,
-						mcuTemp: response.mcutemp ? {
-							min: response.mcutemp.min,
-							current: response.mcutemp.cur,
-							max: response.mcutemp.max
-						} : {},
-						shortName: this.boardType,
-						supports12864: boardDefinition.supports12864,
-						v12: response.v12 ? {
-							min: response.v12.min,
-							current: response.v12.cur,
-							max: response.v12.max
-						} : {},
-						vIn: response.vin ? {
-							min: response.vin.min,
-							current: response.vin.cur,
-							max: response.vin.max
-						} : {}
-					}
-				],
-				fans: newData.fans.map((fanData, index) => ({
-					name: !response.params.fanNames ? null : response.params.fanNames[index],
-					thermostatic: {
-						heaters: ((response.controllableFans & (1 << index)) !== 0) ? [] : [-1]
-					}
-				})),
-				heat: {
-					bedHeaters: response.temps.bed ? [response.temps.bed.heater] : [],
-					chamberHeaters: (response.temps.chamber ? [response.temps.chamber.heater] : [])
-									.concat(response.temps.cabinet ? [response.temps.cabinet.heater] : []),
-					coldExtrudeTemperature: response.coldExtrudeTemp,
-					coldRetractTemperature: response.coldRetractTemp
-				},
-				move: {
-					axes: response.coords.xyz.map((position, index) => ({
-						letter: axisNames[index],
-						visible: (response.axes !== undefined) ? (index < response.axes) : true
-					})),
-					compensation: {
-						type: response.compensation
-					},
-					kinematics: {
-						name: response.geometry
-					}
-				},
-				network: {
-					name: response.name
-				},
-				sensors: {
-					endstops: [...Array(response.coords.xyz.length + response.coords.extr.length)].map((dummy, drive) => ({
-						triggered: Boolean(response.endstops & (1 << drive))
-					})),
-					probes: (response.probe && response.probe.type !== 0) ? [
-						{
-							threshold: response.probe.threshold,
-							triggerHeight: response.probe.height,
-							type: response.probe.type
-						}
-					] : []
-				},
-				state: {
-					machineMode: response.mode ? response.mode : null,
-				},
-				tools: (response.tools !== undefined) ? response.tools.map(tool => ({
-					number: tool.number,
-					name: tool.name ? tool.name : '',
-					heaters: tool.heaters,
-					extruders: tool.drives,
-					axes: tool.axisMap,
-					fans: bitmapToArray(tool.fans),
-					filamentExtruder: (tool.drives.length > 0) ? tool.drives[0] : -1,
-					spindle: tool.spindle,
-					spindleRpm: tool.spindleRpm,
-					offsets: tool.offsets
-				})) : []
-			});
-
-			newData.heat.heaters.forEach(heater => heater.max = response.tempLimit);
-
-			response.tools.forEach(tool => {
-				if (tool.drives.length > 0) {
-					const drive = tool.drives[0];
-					if (drive >= 0 && drive < newData.move.extruders.length) {
-						newData.move.extruders[0].filament = tool.filament;
-					}
-				}
-			});
-
-			if (response.temps.names !== undefined) {
-				response.temps.names.forEach((name, index) => newData.sensors.analog[index].name = name);
-			}
-
-			newData.volumes = [];
-			for (let i = 0; i < response.volumes; i++) {
-				newData.volumes.push({
-					mounted: (response.mountedVolumes & (1 << i)) !== 0
-				});
-			}
-
-			response.tools.forEach(tool => {
-				if (tool.drives.length > 0) {
-					const extruder = tool.drives[0];
-					if (extruder >= 0 && extruder < newData.move.extruders.length && newData.move.extruders[extruder] !== null) {
-						newData.move.extruders[extruder].filament = tool.filament;
-					}
-				}
-			});
-		} else if (statusType === 3) {
-			if (!newData.job) {
-				newData.job = {};
-			}
-
-			// Print Status Response
-			quickPatch(newData.job, {
-				file: {},
-				filePosition: response.filePosition
-			});
-			response.extrRaw.forEach((rawPosition, extruder) => newData.move.extruders[extruder].rawPosition = rawPosition);
-
-			// Update some stats only if the print is still live
-			if (isPrinting) {
-				quickPatch(newData.job, {
-					duration: response.printDuration,
-					layer: response.currentLayer,
-					layerTime: response.currentLayerTime,
-					warmUpDuration: response.warmUpDuration,
-					timesLeft: {
-						file: response.timesLeft.file,
-						filament: response.timesLeft.filament,
-						layer: response.timesLeft.layer
-					}
-				});
-			} else {
-				quickPatch(newData.job, {
-					layerTime: null,
-					timesLeft: {
-						file: null,
-						filament: null,
-						layer: null
-					}
-				});
-			}
-
-			// See if we need to record more layer stats
-			if (response.currentLayer > this.layers.length) {
-				let addLayers = false;
-				if (!this.layers.length) {
-					// Is the first layer complete?
-					if (response.currentLayer > 1) {
-						this.layers.push(new Layer({
-							duration: response.firstLayerDuration,
-							height: this.currentFileInfo.firstLayerHeight || 0,
-							filament: (response.currentLayer === 2) ? response.extrRaw.filter(amount => amount > 0) : [],
-							fractionPrinted: (response.currentLayer === 2) ? response.fractionPrinted / 100 : null
-						}));
-						newData.job.layers = this.layers;
-
-						// Keep track of the past layer
-						if (response.currentLayer === 2) {
-							this.printStats.duration = response.warmUpDuration + response.firstLayerDuration;
-							this.printStats.extrRaw = response.extrRaw;
-							this.printStats.fractionPrinted = response.fractionPrinted;
-							this.printStats.measuredLayerHeight = response.coords.xyz[2] - this.currentFileInfo.firstLayerHeight;
-							this.printStats.zPosition = response.coords.xyz[2];
-						} else if (response.currentLayer > this.layers.length + 1) {
-							addLayers = true;
-						}
-					}
-				} else if (response.currentLayer > this.layers.length + 1) {
-					// Another layer is complete, add it
-					addLayers = true;
-				}
-
-				if (addLayers) {
-					const layerJustChanged = (response.status === 'M') || ((response.currentLayer - this.layers.length) === 2);
-					if (this.printStats.duration) {
-						// Got info about the past layer, add what we know
-						this.layers.push({
-							duration: response.printDuration - this.printStats.duration,
-							height: this.printStats.measuredLayerHeight || this.printStats.layerHeight || 0,
-							filament: response.extrRaw.map((amount, index) => amount - this.printStats.extrRaw[index]).filter((dummy, index) => response.extrRaw[index] > 0),
-							fractionPrinted: (response.fractionPrinted - this.printStats.fractionPrinted) / 100
-						});
-					} else {
-						// Interpolate data...
-						const avgDuration = (response.printDuration - response.warmUpDuration - response.firstLayerDuration - response.currentLayerTime) / (response.currentLayer - 2);
-						for (let layer = this.layers.length; layer + 1 < response.currentLayer; layer++) {
-							this.layers.push(new Layer({
-								duration: avgDuration
-							}));
-						}
-						this.printStats.zPosition = response.coords.xyz[2];
-					}
-					newData.job.layers = this.layers;
-
-					// Keep track of the past layer if the layer change just happened
-					if (layerJustChanged) {
-						this.printStats.duration = response.printDuration;
-						this.printStats.extrRaw = response.extrRaw;
-						this.printStats.fractionPrinted = response.fractionPrinted;
-						this.printStats.measuredLayerHeight = response.coords.xyz[2] - this.printStats.zPosition;
-						this.printStats.zPosition = response.coords.xyz[2];
-					}
-				}
-			}
-		}
-
-		// Output Utilities
-		let beepFrequency = 0, beepDuration = 0, displayMessage = '', messageBox = null;
-		if (response.output) {
-			// Beep
-			if (response.output.beepFrequency !== undefined && response.output.beepDuration !== undefined) {
-				beepFrequency = response.output.beepFrequency;
-				beepDuration = response.output.beepDuration;
-			}
-
-			// Persistent Message
-			if (response.output.message !== undefined) {
-				displayMessage = response.output.message;
-			}
-
-			// Message Box
-			if (response.output.msgBox) {
-				messageBox = new MessageBox({
-					mode: response.output.msgBox.mode,
-					title: response.output.msgBox.title,
-					message: response.output.msgBox.msg,
-					timeout: response.output.msgBox.timeout,
-					axisControls: response.output.msgBox.controls
-				});
-			}
-		}
-
-		quickPatch(newData, {
-			state: {
-				beep: (beepFrequency > 0 && beepDuration > 0) ? new BeepRequest({
-					frequency: beepFrequency,
-					duration: beepDuration
-				}) : null,
-				displayMessage,
-				messageBox
-			}
-		});
-
-		// Spindles
-		if (response.spindles) {
-			quickPatch(newData, {
-				spindles: response.spindles.map(spindle => ({
-					active: spindle.active,
-					current: spindle.current,
-				}))
-			});
-		}
-
-		// Remove invalid heaters and sensors
-		for (let i = 0; i < newData.heat.heaters.length; i++) {
-			const heater = newData.heat.heaters[i];
-			if (heater && heater.state === HeaterState.off) {
-				if (heater.sensor < 0 || heater.sensor >= newData.sensors.analog.length ||
-					newData.sensors.analog[heater.sensor].lastReading === 2000) {
-					newData.heat.heaters[i] = null;
-					newData.sensors.analog[heater.sensor] = null;
-				}
-			}
-		}
-
-		// Update the data model
-		await this.dispatch('update', newData);
-
-		// Check if the G-code response needs to be polled
-		if (response.seq !== this.lastSeq) {
-			await this.getGCodeReply(response.seq);
-		}
-
-		// Check if the firmware rebooted
-		if (!this.justConnected && response.time < this.lastStatusResponse.time) {
-			this.pendingCodes.forEach(code => code.reject(new OperationCancelledError()));
-			this.pendingCodes = [];
-		}
-
-		// Sometimes we need to update the config as well
-		if (requestExtendedStatus || this.justConnected || (this.verbose && this.updateLoopCounter % 2 === 0)) {
-			await this.getConfigResponse();
-		}
-
-		// Status update complete
-		this.lastStatusResponse = response;
-		this.justConnected = false;
-		this.updateLoopCounter++;
-
-		// Schedule the next status update
-		this.scheduleUpdate();
-	}
-
-	convertHeaterState(state) {
-		const keys = Object.keys(HeaterState);
-		if (state >= 0 && state < keys.length) {
-			return keys[state];
-		}
-		return null;
-	}
-
-	convertStatusLetter(letter) {
-		switch (letter) {
-			case 'F': return 'updating';
-			case 'O': return 'off';
-			case 'H': return 'halted';
-			case 'D': return 'pausing';
-			case 'S': return 'paused';
-			case 'R': return 'resuming';
-			case 'P': return 'processing';
-			case 'M': return 'simulating';
-			case 'B': return 'busy';
-			case 'T': return 'changingTool';
-			case 'I': return 'idle';
-		}
-		return 'unknown';
-	}
-
-	// Call this only from updateLoopStatus()
-	async getConfigResponse() {
-		const response = await this.request('GET', 'rr_config');
-		const configData = {
-			boards: [
-				{
-					name: response.firmwareElectronics,
-					firmwareDate: response.firmwareDate,
-					firmwareName: response.firmwareName,
-					firmwareVersion: response.firmwareVersion
-				}
-			],
-			directories: (response.sysdir !== undefined) ? {
-				system: response.sysdir
-			} : {},
-			move: {
-				axes: response.axisMins.map((min, index) => ({
-					acceleration: response.accelerations[index],
-					current: response.currents[index],
-					jerk: response.minFeedrates[index],
-					min,
-					max: response.axisMaxes[index],
-					speed: response.maxFeedrates[index]
-				})),
-				extruders: [...Array(this.numExtruders)].map((dummy, index) => ({
-					acceleration: response.accelerations[response.axisMins.length + index],
-					current: response.currents[response.axisMins.length + index],
-					jerk: response.minFeedrates[response.axisMins.length + index],
-					speed: response.maxFeedrates[response.axisMins.length + index]
-				})),
-				idle: {
-					factor: response.idleCurrentFactor,
-					timeout: response.idleTimeout
-				}
-			},
-			network: {
-				interfaces: [
-					{
-						type: (response.dwsVersion !== undefined) ? 'wifi' : 'lan',
-						firmwareVersion: response.dwsVersion
-					}
-				]
-			}
-		};
-
-		await this.dispatch('update', configData);
-	}
 
 	lastSeqs = {}
 	lastStatus = null
 	lastUptime = 0
+	layers = []
 	webDirectory = Path.web
+
+	pendingCodes = []
 
 	async updateLoopModel() {
 		let jobKey = null, axes = [], extruders = [], analogSensors = [], status = null;
@@ -813,14 +297,8 @@ export default class PollConnector extends BaseConnector {
 
 			// Query the seqs field and the G-code reply
 			this.lastSeqs = (await this.request('GET', 'rr_model', { key: 'seqs' })).result;
-			if (this.lastSeqs == null || this.lastSeqs.reply === undefined) {
-				console.warn('Incompatible rr_model version detected, falling back to status responses');
-				window.forceLegacyConnect = true;
-				BaseConnector.setConnectingProgress(-1);
-				return;
-			}
 			if (this.lastSeqs.reply > 0) {
-				await this.getGCodeReply(this.lastSeqs.reply);
+				await this.getGCodeReply();
 			}
 
 			// Query the full object model initially
@@ -928,10 +406,11 @@ export default class PollConnector extends BaseConnector {
 			this.lastUptime = response.result.state.upTime;
 
 			// Finally check if there is a new G-code reply available
-			if (this.lastSeqs.reply !== seqs.reply) {
-				await this.getGCodeReply(seqs.reply);
-			}
+			const fetchGCodeReply = (this.lastSeqs.reply !== seqs.reply);
 			this.lastSeqs = seqs;
+			if (fetchGCodeReply) {
+				await this.getGCodeReply();
+			}
 		}
 
 		// See if we need to record more layer stats
@@ -1056,13 +535,8 @@ export default class PollConnector extends BaseConnector {
 	async doUpdate() {
 		this.updateLoopTimer = null;
 		try {
-			if (!window.forceLegacyConnect && this.apiLevel >= 1) {
-				// Request object model updates
-				await this.updateLoopModel();
-			} else {
-				// Request updates using the legacy poll adapter
-				await this.updateLoopStatus();
-			}
+			// Request object model updates
+			await this.updateLoopModel();
 		} catch (e) {
 			console.warn(e);
 			await this.dispatch('onConnectionError', e);
@@ -1100,8 +574,8 @@ export default class PollConnector extends BaseConnector {
 		// Check if a response can be expected
 		let result = '';
 		if (strippedCode !== '' && strippedCode.toUpperCase().indexOf('M997') === -1 && strippedCode.toUpperCase().indexOf('M999') === -1) {
-			const pendingCodes = this.pendingCodes, codeSeq = this.lastSeq;
-			result = new Promise((resolve, reject) => pendingCodes.push({ seq: codeSeq, resolve, reject }));
+			const pendingCodes = this.pendingCodes, seq = this.lastSeqs.reply;
+			result = new Promise((resolve, reject) => pendingCodes.push({ seq, resolve, reject }));
 		}
 
 		// Send the code to RRF
@@ -1121,15 +595,12 @@ export default class PollConnector extends BaseConnector {
 		return result;
 	}
 
-	async getGCodeReply(seq = null) {
-		if (seq !== null) {
-			this.lastSeq = seq;
-		}
-
+	async getGCodeReply() {
 		const response = await this.request('GET', 'rr_reply', this.requestTimeout, 'text');
 		const reply = response.trim();
 		if (this.pendingCodes.length > 0) {
 			// Resolve pending code promises
+			const seq = this.lastSeqs.reply;
 			this.pendingCodes.forEach(function(code) {
 				if (seq === null || code.seq < seq) {
 					code.resolve(reply);